refactor(skills): house helper .py under scripts/ and split data.py
Standardize the skills/ pipeline on the same scripts/ layout the older
tableau-dashboard-creator skill (and the write-a-skill guideline) already
use, and break up the overloaded tableau-data helper.

File Organization
- Move route.py, init.py, intake.py, data.py, vds.py from each skill root
  into skills/<skill>/scripts/ (git mv, history preserved).
- Fix __file__-relative resource lookups now one level deeper: init.py's
  skeleton/ and intake.py's references/PRD-TEMPLATE.md -> .parent.parent.
- Update the four SKILL.md invocation paths, three CONTRACT.md source
  pointers, and tests/conftest.py sys.path to the scripts/ subdirs.

Enhancements
- Split data.py (1528 lines, ~6 jobs) into focused modules:
  - constants.py  (66)  - CONTRACT.md-mirror constants (leaf, breaks cycle)
  - datamodel.py  (477) - type inference, CSV profiling, DATA-MODEL.md
                          render/parse, header validation
  - state.py      (201) - STATE.md parse/rewrite, data_mode, staleness, gate
  - data.py       (893) - command layer (precheck/profile/pull/commit) + CLI
- data.py re-exports the names the contract test uses, so the public
  data.* surface is unchanged (no test churn).
- Promote three helpers from _private to public (suggest_role,
  distinct_samples, downstream_stale_updates) now that a sibling module
  legitimately imports them.

Validation
- Full suite green (113 passed), unchanged from baseline.
- All modules py_compile; no stale path references repo-wide.
- Live CLI runs: init scaffold, intake template render, route, and the
  data precheck -> profile -> commit csv flow.
